fix failing assertions not displayed in test output
A failing expectation throws an error and the callback therefore never
reaches done(). This then leads to a timeout error and the actual failure
is never displayed. This can be solved by wrapping all expectations
in try/catch and then calling done from within try and from catch including
the error. I choose to simply modernize all tests to use promises which
can be simply returned from the test case and are handled in a more elegant
way by mocha itself.